ECO-2286 update of templates in order to use latest Atomic design
