--- src/SprykerEco/Yves/AmazonPay/Theme/default/widget/wallet-widget.twig
@@ -1,10 +1,0 @@
-<div id="amazonpayWalletWidgetContainer"> </div>
-
-<script>
-    var amazonpayConfig = {
-        clientId: "{{ amazonpayConfig.clientId }}",
-        sellerId: "{{ amazonpayConfig.sellerId }}"
-    };
-</script>
-
-<script src="/assets/default/js/spryker-yves-amazon-pay-wallet.js"></script>
